+++ internal/usecase/interface.go
@@ -10,6 +10,6 @@ type TaskListRepo interface {
 	DeleteTask(ctx context.Context, id int64) error
 	Update(ctx context.Context, t *entity.Task) error
 	GetAllTasks(ctx context.Context) ([]entity.Task, error)
-	GetTaskById(ctx context.Context, id int) (*entity.Task, error)
+	GetTaskById(ctx context.Context, id int64) (*entity.Task, error)
 	GetAllTasksPages(ctx context.Context, n int) (map[int][]entity.Task, error)
 }
